backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
websockets==13.1

# Google Cloud & Vertex AI
google-genai==1.56.0
google-cloud-aiplatform==1.73.0
google-auth==2.36.0
google-api-core==2.23.0

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.11.4

# Utilities
python-dotenv==1.0.1
pydantic==2.10.2
pydantic-settings==2.6.1

# Async & Concurrency
aiofiles==24.1.0
asyncio==3.4.3

# Optional: Redis for production
redis==5.2.0
aioredis==2.0.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.27.2
